requirements.txt
requests
aiohttp
pacifica-sdk
python-dotenv
python-telegram-bot==21.4
